requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
mistralai>=0.0.7
aiohttp>=3.8.0
